\docType{data}
\name{ameshousing}
\alias{ameshousing}
\title{Ames housing statistics}
\format{data.frame: 1615 obs. of  10 variables}
\source{
  \url{http://www.cityofames.org/Assessor/index.htm}
}
\description{
  This data contains statistics for the housing market in
  Ames, Iowa from January 2008 to September 2012
}
\examples{
data(ameshousing)
summary(ameshousing)
}
\keyword{datasets}
