\name{bprint}
\alias{bprint}
\title{Print Item Information..}
\usage{bprint(...)

}
\description{Print Item Information
prints the item name and structure information below it}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{hplot}
\arguments{\item{...}{item to be quoted and then printed}
}
\examples{dataRanges <- c(0, 1, 2, 3)
bprint(dataRanges)}

