\name{common_key_press}
\alias{common_key_press}
\alias{common_key_release}
\title{Some common processings in the key press and release events}
\usage{
  common_key_press(layer, event, data, meta)

  common_key_release(layer, event, data, meta)
}
\arguments{
  \item{layer}{the layer argument in the event callback}

  \item{event}{the event argument in the event callback}

  \item{data}{the data created by \code{\link{qdata}}}

  \item{meta}{the meta data for a plot}
}
\value{
  \code{NULL}
}
\description{
  The key press and release events often involve with
  setting the selection mode of the \code{\link{brush}},
  the alpha transparency, and deleting selected elements,
  and so on. These functions implement these common
  processes.

  In a key release event, we set the selection mode to
  \code{'none'}. If PageUp or PageDown was pressed, we show
  the brush history.
}
\details{
  The keys A, O, X, N and C corresponds to the selection
  mode AND, OR, XOR, NOT and COMPLEMENT respectively.

  Plus (+) and Minus (-) can increase or decrease the alpha
  transparency exponentially.

  The key Delete will make the selected elements invisible,
  and F5 makes all the elements visible.

  The question key (?) toggles the identify mode (on or
  off). In the identify mode, labels generated by the
  identified cases will be drawn under the mouse. The cross
  cursor shape (+) indicates it is in the identify mode,
  and a normal cursor indicates the brush mode.
}
\examples{
## see the source code of qbar() or qparallel()
}
\author{
  Yihui Xie <\url{http://yihui.name}>
}
\seealso{
  \code{\link{brush}}
}
