\name{draw_identify}
\alias{draw_identify}
\title{Draw the identify labels under the mouse}
\usage{
draw_identify(layer, painter, data, meta)
}
\arguments{
  \item{layer}{the layer argument of the painting function}

  \item{painter}{the painter argument of the painting
  function}

  \item{data}{the data object used by the plot}

  \item{meta}{the meta object in the plot}
}
\value{
\code{NULL}
}
\description{
The identify labels are generated by specific plotting functions, and this
function draw them as texts under the mouse, with a transparent gray
background.
}
\details{
These components in the meta data are required to draw the labels:
\code{meta$pos} (mouse position), \code{meta$limits} (layer limits) and
\code{meta$identify.labels} (labels to draw). The color of the labels is set
in the \code{\link{brush}} object as \code{brush(data)$label.color}.

The label will automatically adjust its own positions to ensure it does not
exceed the plot margin. For example, when the mouse is near the right margin,
the label will be drawn to the left.
}
\examples{
## see source code of qparallel() or qhist()
}
\author{
Yihui Xie <\url{http://yihui.name}>
}
