\name{draw_x_axes_with_labels}
\alias{draw_x_axes_with_labels}
\title{draw x axes...}
\usage{draw_x_axes_with_labels(plotObj, dataRanges, axisLabels, labelHoriPos, name, row = 0, 
    col = 0)
}
\description{draw x axes
draws the x axes with the labels and label positions given}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{hplot}
\arguments{\item{plotObj}{Qt plot object to have the layer added to}
\item{dataRanges}{ranges of the data so a buffer of space may be added}
\item{axisLabels}{vector of labels}
\item{labelHoriPos}{horizontal position of the axisLabels}
}
