\name{draw_y_axes_with_labels}
\alias{draw_y_axes_with_labels}
\title{draw y axes...}
\usage{draw_y_axes_with_labels(plotObj, dataRanges, axisLabels, 
    labelVertPos, name, row = 0, col = 0)

}
\description{draw y axes
draws the y axes with the labels and label positions given}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{hplot}
\arguments{\item{plotObj}{Qt plot object to have the layer added to}
\item{dataRanges}{ranges of the data so a buffer of space may be added}
\item{axisLabels}{vector of labels}
\item{labelVertPos}{vertical position of the axisLabels}
}
