% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lighter}
\alias{lighter}
\title{Generate lighter colors}
\usage{
lighter(color, factor = 0.2)
}
\arguments{
\item{color}{the color vector}

\item{factor}{numeric: larger than 0 means to make the color lighter, and
less than 0 means darker (can be either a vector or a scalar)}
}
\value{
The modified color vector.
}
\description{
This function first converts colors to the RGB and HSV space, then modifies
the brightness (the \code{V} component in HSV) by an amount.
}
\examples{
lighter(palette())

par(mfrow = c(2, 1), mar = rep(0, 4))
d <- c("black", grep("dark", colors(), value = TRUE))  # dark colors
x <- rep(1, length(d))
for (f in seq(0, 1, length = 20)) {
    barplot(x, col = d)
    barplot(x, col = lighter(d, f))
    if (interactive()) 
        Sys.sleep(0.2)
}
}
\author{
Heike Hofmann and Yihui Xie
}
