% Generated by roxygen2 (4.0.1): do not edit by hand
\name{link_id}
\alias{link_id}
\title{Get the id's of listeners on linking}
\usage{
link_id(data)
}
\arguments{
\item{data}{a mutaframe created by \code{\link{qdata}}}
}
\value{
A character vector of id's.
}
\description{
This function returns the id's of listeners attached on a data for linking,
and these id's can be used to remove the linking.
}
\examples{
## see ?link_cat
}
\author{
Yihui Xie <\url{http://yihui.name}>
}
