\name{make_data_ranges}
\alias{make_data_ranges}
\title{makes the data ranges...}
\usage{make_data_ranges(dataColumn)
}
\description{makes the data ranges
makes the data ranges with a small buffer}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{hplot}
\arguments{\item{dataColumn}{vector containing numeric elements}
}
\examples{make_data_ranges(mtcars$disp)}
