\name{make_dodge_pos}
\alias{make_dodge_pos}
\title{Make dodge positions}
\usage{
  make_dodge_pos(breaks, n)
}
\arguments{
  \item{breaks}{break positions}

  \item{n}{number of items per break}
}
\description{
  Make dodge positions
}
\examples{
make_dodge_pos(c(1:5), 3)
}
\author{
  Barret Schloerke \email{bigbear@iastate.edu}
}
\keyword{internal}
