\name{make_new_plot}
\alias{make_new_plot}
\title{make a plot object...}
\usage{make_new_plot(windowRanges, width = 600, height = 600)

}
\description{make a plot object
make a plot object}
\value{Returns a Qt plot object}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{hplot}
\arguments{\item{windowRanges}{Ranges of the plotting window}
}
\examples{make_new_plot(make_window_ranges(c(0, 1, 2, 3)))}
