\name{make_pretty_axes}
\alias{make_pretty_axes}
\title{pretty axes...}
\usage{make_pretty_axes(dataRange, minimum, maximum)

}
\description{pretty axes
make 'pretty' axes and make sure that they are within the minimum and maximum}
\arguments{\item{dataRange}{data range that should be 'pretty'}
\item{minimum}{minimum value of the 'pretty' range}
\item{maximum}{maximum value of the 'pretty' range}
}
