\name{nrcstat}
\alias{nrcstat}
\alias{nrcstat}
\title{NRC Rankings Data for the Statistics Departments in US.}
\description{NRC Rankings Data for the Statistics Departments in US.
This data contains the NRC rankings for all the statistics departments in US.}
\docType{data}
\keyword{datasets}
\examples{data(nrcstat)
summary(nrcstat)}

