\name{qmap}
\alias{qmap}
\title{Interactive Maps}
\usage{qmap(data, longitude, latitude, group, by.x, label = group, 
    labeldata, by.y = by.x, colour, main, ...)

}

\description{
Create an interactive map from qdata}
\value{NULL}
\author{Heike Hofmann}
\arguments{\item{data}{a mutaframe which is typically built upon a data frame
along with several row attributes}
\item{latitude}{spatial x variable}
\item{longitude}{spatial y variable}
\item{group}{grouping variable for polygons}
\item{colour}{fill colour of the polygons}
\item{...}{}
}
