\name{qtour}
\alias{qtour}
\title{Create a tour associated with a mutaframe}
\usage{
qtour(vars = ~., data = last_data(), tour_path = grand_tour(), 
    aps = 1, fps = 30, rescale = TRUE, sphere = FALSE, ...)
}
\arguments{
  \item{vars}{variable names to be used in the tour (parsed
  by \code{\link{var_names}})}

  \item{data}{a mutaframe created by \code{\link{qdata}}
  (default to be \code{\link{last_data}()}, i.e. the lastly
  used data)}

  \item{tour_path}{tour path generator, defaults to 2d
  grand tour}

  \item{aps}{target angular velocity (in radians per
  second)}

  \item{fps}{target frames per second (defaults to 30)}

  \item{rescale}{if true, rescale all variables to range
  [0,1]?}

  \item{sphere}{if true, sphere all variables}

  \item{...}{ignored}
}
\value{
  An object generated by reference classes with signals.
}
\description{
  The \pkg{tourr} package is used to create the tour
  projections. This function creates an \R object to
  manipulate the tour, and all the changes in the tour can
  be reflected immediately in plots created in
  \pkg{cranvas}.
}
\details{
  Because the data provided to the tour is a mutaframe, it
  can listen to changes through listeners on it. Usually
  these listeners can update the plots which are created
  from this mutaframe (further, those linked to this
  mutaframe) on the fly.

  Four basic methods can be applied to the object returned
  by this function (say, \code{tour}): \code{tour$start()}
  will start the tour (tour projections change successively
  and are attached to the mutaframe; column names are
  \code{tour_1}, \code{tour_2}, ...); \code{tour$stop()}
  pauses the tour; \code{tour$slower()} makes the tour
  slower and \code{tour$faster()} makes it faster.
}
\examples{
## examples of tourrs in cranvas
library(cranvas)
data(flea, package = "tourr")

qflea <- qdata(flea, color = species)
flea_tour <- qtour(1:6, data = qflea, tour_path = grand_tour(3))
flea_tour$start()

qparallel(~tour_1 + tour_2 + tour_3)
qhist(tour_1, binwidth = 0.05, xlim = c(-1, 1), ylim = c(0, 
    16))

flea_tour$pause()

flea_tour$slower()

flea_tour$faster()
}
\author{
  Yihui Xie <\url{http://yihui.name}>
}
