\name{switch_serie_mode}
\alias{switch_serie_mode}
\title{key M for switching the serie mode
on the serie mode users can drag any serie horizontally}
\usage{
switch_serie_mode(meta, data)
}
\description{
  key M for switching the serie mode on the serie mode
  users can drag any serie horizontally
}
