\name{zero_then_top_by_order}
\alias{zero_then_top_by_order}
\title{Collect 0 and ordered vector...}
\usage{zero_then_top_by_order(vec)

}
\description{Collect 0 and ordered vector}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\keyword{internal}
\arguments{\item{vect}{vector to be used}
\item{vect_order}{order to be used (except for the last item)}
}
