\name{make_dodge_pos}
\alias{make_dodge_pos}
\title{Make dodge positions..}
\usage{make_dodge_pos(breaks, n)

}
\description{Make dodge positions}
\keyword{internal}
\author{Barret Schloerke \email{bigbear@iastate.edu}}
\arguments{\item{breaks}{break positions}
\item{n}{number of items per break}
}
\examples{make_dodge_pos(c(1:5), 3)}
