\docType{data}
\name{nrcstat}
\alias{nrcstat}
\title{NRC Rankings Data for the Statistics Departments in US}
\format{data.frame: 61 obs. of  68 variables}
\source{
  \url{http://sites.nationalacademies.org/pga/resdoc/index.htm}
}
\description{
  This data contains the NRC rankings for all the
  statistics departments in US.
}
\examples{
data(nrcstat)
summary(nrcstat)
}
\keyword{datasets}
