\name{qscatter}
\alias{qscatter}
\title{Draw a scatterplot..}
\usage{qscatter(data, form, main, labeled = TRUE)

}
\description{Draw a scatterplot}
\arguments{\item{data}{data.frame source}
\item{form}{formula in format y ~x which designates the axis}
\item{main}{main title for the plot}
\item{labeled}{whether axes should be labeled}
}
