\name{qscatter}
\alias{qscatter}
\title{Draw a scatter plot}
\usage{
qscatter(x, y, data = last_data(), main = "", xlim = NULL, 
    ylim = NULL, xlab = NULL, ylab = NULL, asp = NULL)
}
\arguments{
  \item{x}{the name of the x variable}

  \item{y}{the name of the y variable (if missing, \code{x}
  will be plotted against its indices, i.e., \code{y} will
  become \code{x} and \code{x} will be indices)}

  \item{asp}{aspect ratio (\code{NULL} means to use the
  data ranges directly, and a positive value of the aspect
  ratio will change the limits of one of the two axes); the
  implementation is not perfect yet, since resizing the
  plotting window will change the aspect ratio (different
  to other R graphics systems)}

  \item{data}{a mutaframe created by \code{\link{qdata}}
  (default to be \code{\link{last_data}()}, i.e. the lastly
  used data)}

  \item{main}{the main title}

  \item{xlim}{a numeric vector of length 2 (like
  \code{c(x0, x1)}) for x-axis limits; it will be
  calculated from the data limits if not specified
  (\code{NULL}). Note when \code{x0 > x1}, the axis
  direction will be reversed (i.e. from larger values to
  small values)}

  \item{ylim}{y-axis limits; similar to \code{xlim}}

  \item{xlab}{x-axis title}

  \item{ylab}{y-axis title}
}
\value{
  A scatter plot
}
\description{
  This function creates a scatter plot with two numeric
  variables. It supports colors and sizes generated by
  \code{\link{qdata}}. When the sizes are not the same, a
  scatter plot becomes what is called the ``bubble chart''.
}
\details{
  All the common interactions like brushing and deleting
  are documented in \code{\link{common_key_press}}.

  In the identify mode, the plot will show the row id(s) of
  the identified point(s) as well as \code{x} and \code{y}
  values.

  Arrow keys are used in scatter plots for interactions:
  without the \code{Shift} key, \code{Up} and \code{Down}
  will increase/decrease the sizes of points; with the
  \code{Shift} key being pressed, \code{Up}, \code{Down},
  \code{Left} and \code{Right} will move the points to the
  corresponding direction.

  Mouse wheel is used to zoom in and zoom out the plot.
  Together with keyboard interactions, this will enable us
  to see details in data.
}
\note{
  Using more than one color (including border colors) can
  lead to serious speed issues; this plot is fastest when
  we only use a single color and a single size. As the
  number of unique colors increases, the plotting can be
  slower based on a same dataset. Using more than one size
  is also slower than using a single size for all points;
  the internal painting functions are different for the two
  cases (\code{\link[qtpaint]{qdrawCircle}} for the former,
  and \code{\link[qtpaint]{qdrawGlyph}} for the latter).
}
\examples{
library(cranvas)

### (1) flea: color by categorical variable, and linking
data(flea, package = "tourr")
qflea <- qdata(flea, color = species)  # use species to create colors

qscatter(tars1, aede1)
qscatter(tars1, tars2, asp = 1)  # aspect ratio

## link qflea to itself using species
id <- link_cat(qflea, "species")


## remove linking
remove_listener(qflea, id)


## a bubble chart
qflea2 <- qdata(flea, color = NA, border = species, 
    size = tars1)
qscatter(tars1, tars2)



### (2) NRC rankings
data(nrcstat)

qnrc <- qdata(nrcstat, color = Regional.Code)

qscatter(R.Rankings.5th.Percentile, R.Rankings.95th.Percentile)

qscatter(S.Rankings.5th.Percentile, S.Rankings.95th.Percentile)



\dontrun{
### (3) secrets in the pollen data
library(animation)
data(pollen, package = "animation")
head(pollen)
qpollen <- qdata(pollen, color = rgb(0, 0, 0, 0.2), size = 2)
qscatter(RIDGE, CRACK, data = qpollen)
## try zooming into the center



### (4) pressure test; run with care!
n <- 1e+06  # a million still works (at least for me)
df <- qdata(data.frame(x = rnorm(n), y = rnorm(n), z = gl(4, 
    n/4)), color = z)
qscatter(x, y)
}
}
\author{
  Heike Hofmann, Di Cook, Yihui Xie
}
