\name{qtime}
\alias{qtime}
\title{Draw a time plot}
\usage{
qtime(time, y, data, period = NULL, group = NULL, shift = c(1, 4, 7, 12, 24), size = 3, 
    alpha = 1, asp = NULL, similarity.index = TRUE, help.text = TRUE, main = NULL, 
    xlab = NULL, ylab = NULL, ...)
}
\arguments{
  \item{time}{the variable indicating time on the
  horizontal axis.}

  \item{y}{a vertor of all the variable names of interest.}

  \item{data}{Mutaframe data generated by
  \code{\link{qdata}}.}

  \item{period}{The variable to group the time series.
  Better to be 'year','month', or other time resolutions.
  Default to be null. When it is not null, the key U and D
  can be hit to separate the groups or overlap them
  together to watch the patterns.}

  \item{group}{Similar to period, but is used for
  longitudinal data grouping.}

  \item{shift}{Wrapping speed selector. The default
  possible speeds are 1,7(for days a week),12(for
  months),24(for hours).}

  \item{size}{Point size, default to be 2.}

  \item{alpha}{Transparency level, 1=completely opaque,
  default to be 1.}

  \item{asp}{Ratio between width and height of the plot.}

  \item{similarity.index}{Whether to show the statistics
  which measure the similarity between series when
  wrapping. It gives the ACF, corr, and R square for one,
  two, and more series respectively.}

  \item{help.text}{Whether to show the instruction. All the
  text will disappear in 2 seconds after the interaction.}

  \item{main}{main title for the plot.}

  \item{xlab}{label on horizontal axis, default is name of
  x variable}

  \item{ylab}{label on vertical axis, default is name of y
  variable}
}
\description{
  Draw a time-series plot.
}
\details{
  Arrow up/down: in-/de-crease size of points. Arrow
  left/right: wrap the time series when wrap=TRUE, while
  zoom in/out with the center of the last clicked dot when
  wrap=FALSE. Shift + right: when wrap=TRUE, the time
  series will be folded directly to the width of maximal
  value in argument shift. Shift + left: time series will
  be backed to the original xaxis position, no matter wrap
  is TRUE or FALSE. Key '+'/'-': de-/in-crease alpha level
  (starts at alpha=1 by default). Key 'u'/'d': separate/mix
  the series groups by shifting them up and down. Shift +
  'u'/'d': for multivariate y's, separate/mix them by
  shifting up and down. Key 'g': change the wrapping period
  circularly in the values of parameter 'shift'. Key 'm':
  Switch the mode for series selecting. Default to be off.
  When there are more than one series in the plot, users
  can turn it on to hold a series and shift the series
  horizontally by dragging with the mouse. Wheel: Zoom
  in/out. Then users can drag the series horizontally to
  see the details.
}
\examples{
library(cranvas)

## example 1: NASA temperature data
nasa2221 <- subset(nasa, Gridx == 22 & Gridy == 21)
nasa2221$Year <- factor(nasa2221$Year)
qnasa <- qdata(nasa2221)

qtime(TimeIndx, ts, qnasa, shift = c(1, 12))
qscatter(ts, ps_tovs, data = qnasa)

qtime("TimeIndx", "ts", qnasa, Year, shift = 1)
qtime(TimeIndx, c("ts", "ps_tovs", "ca_med"), qnasa, shift = c(1, 12))
qtime("TimeIndx", c(ts, ps_tovs, ca_med), qnasa, period = Year)


## example 2: Remifentanil in the nlme package
library(nlme)
qRem <- qdata(Remifentanil[complete.cases(Remifentanil) & Remifentanil$ID == 1, ])
qtime(Time, conc, qRem)

Remi <- Remifentanil[complete.cases(Remifentanil), ]
Remi$ID <- factor(Remi$ID)
qRemi <- qdata(Remi)
qtime(Time, conc, qRemi, group = ID)
# for categorical brushing self-link dataset by ID: id <- link_cat(qRemi, 'ID')
# remove_link(qRemi, id)


## example 3: Wages
qwage <- qdata(wages[as.integer(as.character(wages$id)) < 2000, 1:3])
qtime(exper, lnw, qwage, group = id)
# id <- link_cat(wage, 'id') remove_link(wage, id)


## example 4: Lynx - for posterity Good to show off wrapping to investigate
## irregular series
qlynx <- qdata(data.frame(Time = 1:114, lynx))
qtime(Time, lynx, qlynx, shift = 1:12)


## example 5: Sunspots - for posterity Good to show off wrapping to investigate
## irregular series
qsun <- qdata(data.frame(Time = 1:2820, sunspots))
qtime(Time, sunspots, qsun, shift = c(1, (1:10) * 10))


## example 6: Pigs
qpig <- qdata(pigs)
qtime(TIME, c("GILTS", "PROFIT", "PRODUCTION", "HERDSZ"), qpig, shift = c(1, 4))

cranvas_off()
}
\seealso{
  Other plots: \code{\link{qbar}}, \code{\link{qboxplot}},
  \code{\link{qdensity}}, \code{\link{qhist}},
  \code{\link{qmval}}, \code{\link{qparallel}},
  \code{\link{qspine}}
}
